Optimized scanimage, revised Mp classes, moved MpSession and MpScan to commons, overhauled opt classes

+++ schemas/+psy/Grating.m
@@ -11,10 +11,10 @@ aperture_x=0                : float                         # aperture x coordin
 aperture_y=0                : float                         # aperture y coordinate, in units of half-diagonal, 0 = center
 grating                     : enum('sqr','sin')             # sinusoidal or square, etc.
 drift_fraction=0            : float                         # the fraction of the trial duration taken by drifting grating
-spatial_freq                : float                         # cycles/degree
+spatial_freq                : decimal(4,2)                  # cycles/degree
 init_phase                  : float                         # 0..1
 trial_duration              : float                         # ms
-temp_freq                   : float                         # Hz
+temp_freq                   : decimal(4,1)                  # Hz
 direction                   : float                         # 0-360 degrees
 %}
